# docs/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Noor API documentation",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "https://wwwwwwww123.herokuapp.com/",
  "basePath": "/",
...
  "schemes": [
    "https",
    "http"
  ],
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "access-token",
      "in": "header"
    },
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "https://petstore.swagger.io/oauth/authorize",
      "flow": "implicit",
      "scopes": {
        "read_pets": "read your pets",
        "write_pets": "modify pets in your account"
      }
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/client/register": {
      "post": {
        "tags": [
          "Client"
        ],
        "description": "client services",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "first_name": {
                  "example": "any"
                },
                "last_name": {
                  "example": "any"
                },
                "father_name": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "img_id": {
                  "example": "any"
                },
                "phone": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "role": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Adding clients",
            "schema": {
              "$ref": "#/definitions/Client"
            }
          },
          "400": {
            "description": "Xatolik roy berdi"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/client/login": {
      "post": {
        "tags": [
          "Client"
        ],
        "description": "client services",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Thsi api helps to login",
            "schema": {
              "$ref": "#/definitions/Client"
            }
          },
          "400": {
            "description": "Xatolik roy berdi"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/client/list": {
      "post": {
        "tags": [
          "Client"
        ],
        "description": "this api response is list of all clients",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "pageNumber": {
                  "example": "any"
                },
                "pageSize": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "client services",
            "schema": {
              "$ref": "#/definitions/Client"
            }
          },
          "400": {
            "description": "Xatolik roy berdi"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/client/resetpasswordclient": {
      "post": {
        "tags": [
          "Client"
        ],
        "description": "this api response is list of all clients",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "client services",
            "schema": {
              "$ref": "#/definitions/Client"
            }
          },
          "400": {
            "description": "Xatolik roy berdi"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/client/resetpassword/confirmationp/{email}/{token}": {
      "post": {
        "tags": [
          "Client"
        ],
        "description": "this api response is list of all clients",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "token",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "client services",
            "schema": {
              "$ref": "#/definitions/Client"
            }
          },
          "400": {
            "description": "Xatolik roy berdi"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/client/update/{id}": {
      "post": {
        "tags": [
          "Client"
        ],
        "description": "client services",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "first_name": {
                  "example": "any"
                },
                "last_name": {
                  "example": "any"
                },
                "father_name": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "img_id": {
                  "example": "any"
                },
                "phone": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "role": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "This api helps to update exact user by id",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Xatolik roy berdi"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/client/auth/{id}": {
      "patch": {
        "tags": [
          "Client"
        ],
        "description": "client services",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "This api helps to update exact user by id",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Xatolik roy berdi"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/client/delete/{id}": {
      "delete": {
        "tags": [
          "Client"
        ],
        "description": "this api helps to delete exact client by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Client successfully deleted",
            "schema": {
              "$ref": "#/definitions/Client"
            }
          },
          "400": {
            "description": "Xatolik roy berdi"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/client/getone/{id}": {
      "get": {
        "tags": [
          "Client"
        ],
        "description": "api hels to get exact client by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "The result of this",
            "schema": {
              "$ref": "#/definitions/Client"
            }
          },
          "400": {
            "description": "Xatolik roy berdi"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/user/register": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "client services",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "first_name": {
                  "example": "any"
                },
                "last_name": {
                  "example": "any"
                },
                "father_name": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "img_id": {
                  "example": "any"
                },
                "phone": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "role": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Adding clients",
            "schema": {
              "$ref": "#/definitions/Client"
            }
          },
          "400": {
            "description": "Xatolik roy berdi"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "client services",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Thsi api helps to login",
            "schema": {
              "$ref": "#/definitions/Client"
            }
          },
          "400": {
            "description": "Xatolik roy berdi"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/user/list": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "this api response is list of all clients",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "pageNumber": {
                  "example": "any"
                },
                "pageSize": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "client services",
            "schema": {
              "$ref": "#/definitions/Client"
            }
          },
          "400": {
            "description": "Xatolik roy berdi"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/user/resetpassworduser": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "this api response is list of all clients",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "client services",
            "schema": {
              "$ref": "#/definitions/Client"
            }
          },
          "400": {
            "description": "Xatolik roy berdi"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/user/resetpassword/confirmationp/{email}/{token}": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "this api response is list of all clients",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "token",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "client services",
            "schema": {
              "$ref": "#/definitions/Client"
            }
          },
          "400": {
            "description": "Xatolik roy berdi"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/user/auth/{id}": {
      "patch": {
        "tags": [
          "User"
        ],
        "description": "client services",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "This api helps to update exact user by id",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Xatolik roy berdi"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Client": {
      "type": "object",
      "properties": {
        "first_name": {
          "type": "string",
          "example": "Doston"
        },
        "last_name": {
          "type": "string",
          "example": "Ashurov"
        },
        "father_name": {
          "type": "string",
          "example": "Baxtiyorovich"
        },
        "email": {
          "type": "string",
          "example": "ashurov@list.ru"
        },
        "img_id": {
          "type": "string",
          "example": "6184b677f552e1ab7baf6fa2"
        },
        "phone": {
          "type": "string",
          "example": "94-456-7891"
        },
        "password": {
          "type": "string",
          "example": "789456qwer#A"
        },
        "role": {
          "type": "string",
          "example": "client"
        }
      }
    },
...
  }
}